% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_fec}
\alias{etl_extract.etl_fec}
\alias{etl_load.etl_fec}
\alias{etl_transform.etl_fec}
\title{ETL operations for FEC data}
\source{
\url{http://www.fec.gov/finance/disclosure/ftpdet.shtml}
}
\usage{
\method{etl_extract}{etl_fec}(obj, years = 2014, ...)

\method{etl_transform}{etl_fec}(obj, years = 2014, ...)

\method{etl_load}{etl_fec}(obj, years = 2014, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{years}{a vector of integers representing the years}

\item{...}{arguments passed to methods}
}
\description{
ETL operations for FEC data
}
\details{
If a \code{year} and/or \code{month} is specified, then
only flight data from matching months is used.
}
\examples{
\dontrun{
fec <- etl("fec", dir = "~/dumps/fec")
fec \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_init() \%>\%
  etl_load()
}
\dontrun{
if (require(RMySQL)) {
  # must have pre-existing database "fec"
  # if not, try
  system("mysql -e 'CREATE DATABASE IF NOT EXISTS fec;'")
  db <- src_mysql(default.file = path.expand("~/.my.cnf"), groups = "rs-dbi",
                  user = NULL, password = NULL, dbname = "fec")
}

fec <- etl("fec", db, dir = "~/dumps/fec")
fec \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_init() \%>\%
  etl_load()
}
}

