% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-package.R
\docType{package}
\name{fec-package}
\alias{fec}
\alias{fec-package}
\title{fec: Free English Classes registration website}
\description{
Free English Classes registration website
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabrielburcea/fec}
  \item Report bugs at \url{https://github.com/gabrielburcea/fec/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Burcea \email{gb.burcea@gmail.com}

}
\keyword{internal}
