% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_student_table}
\alias{get_student_table}
\title{Collect student data}
\usage{
get_student_table(db_path = app_sys("fec_data.sqlite"))
}
\arguments{
\item{db_path}{Path to the database.}
}
\value{
A data frame.
}
\description{
Collect student data from the database
}
