% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_teacher_table}
\alias{get_teacher_table}
\title{Collect teacher data}
\usage{
get_teacher_table(db_path = app_sys("fec_data.sqlite"))
}
\arguments{
\item{db_path}{Path to the database.}
}
\value{
A data frame.
}
\description{
Collect teacher data from the database.
}
