% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{simplify_string}
\alias{simplify_string}
\title{helpers}
\usage{
simplify_string(x)
}
\arguments{
\item{x}{a character vector or a factor. Factors will be silently converted
to character strings.}
}
\value{
The return value, if any, from executing the utility.

character vector of the same length as the input vector.
}
\description{
A utils function
}
\examples{
simplify_string(c(" Dirty_name. to Clean.#$", "#$another   complicated. Name"))
}
