% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{teaching_levels}
\alias{teaching_levels}
\title{Teaching levels}
\format{
An object of class \code{character} of length 3.
}
\usage{
teaching_levels
}
\description{
A character vector with the teaching levels available.
}
\keyword{datasets}
